src/services/wisdom_service/requirements.txt
flask==2.2.3
requests==2.28.2
python_dotenv==1.0.0